ψ/lab/thai-legal-rag/requirements.txt
# Thai Legal RAG v2 — pinned dependencies
# Last updated: 2026-02-13

# Google
google-api-python-client==2.156.0
google-auth==2.37.0
google-generativeai==0.8.4

# PDF processing
PyMuPDF==1.25.2       # fitz: PDF → images for OCR

# Vector store
faiss-cpu==1.9.0
numpy==1.26.4

# Graph RAG
lightrag-hku==1.3.1

# Thai NLP
pythainlp==5.0.5

# ML utilities
sentence-transformers==3.4.1

# Pipeline
tqdm==4.67.1
python-dotenv==1.0.1

# Dedup DB
# sqlite3 is stdlib — no extra dep needed

# UI
streamlit==1.42.0

# Tests
pytest==8.3.4
pytest-asyncio==0.25.3
